Wrote a program that list the fisrt fifty fibonacci number

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@ int main(void)
 {
 int n = 50;
 int i;
-unsigned long long int nf1 = 1;
-unsigned long long int nf2 = 2;
-printf("%llu, %llu, ", nf1, nf2);
+long int nf1 = 1;
+long int nf2 = 2;
+printf("%ld, %ld, ", nf1, nf2);
 for (i = 3; i <= n; i++)
 {
-unsigned long long int nnf = nf1 + nf2;
-printf("%llu", nnf);
+long int nnf = nf1 + nf2;
+printf("%ld", nnf);
 if (i != n)
 {
 printf(", ");
